Add step and repeat support to Parser2

+++ src/pygerber/gerberx3/tokenizer/tokens/sr_step_repeat.py
@@ -1,6 +1,7 @@
 """Wrapper for aperture select token."""
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from pygerber.gerberx3.tokenizer.tokens.bases.command import CommandToken
@@ -24,10 +25,10 @@ class StepRepeatBegin(CommandToken):
         self,
         string: str,
         location: int,
-        x_repeat: float,
-        y_repeat: float,
-        x_step: float,
-        y_step: float,
+        x_repeat: int,
+        y_repeat: int,
+        x_step: Decimal,
+        y_step: Decimal,
     ) -> None:
         super().__init__(string, location)
         self.x_repeat = x_repeat
@@ -41,10 +42,10 @@ class StepRepeatBegin(CommandToken):
 
         Created to be used as callback in `ParserElement.set_parse_action()`.
         """
-        x_repeat = float(str(tokens.get("x_repeat", "0")))
-        y_repeat = float(str(tokens.get("y_repeat", "0")))
-        x_step = float(str(tokens.get("x_step", "0")))
-        y_step = float(str(tokens.get("y_step", "0")))
+        x_repeat = int(str(tokens.get("x_repeat", "0")))
+        y_repeat = int(str(tokens.get("y_repeat", "0")))
+        x_step = Decimal(str(tokens.get("x_step", "0")))
+        y_step = Decimal(str(tokens.get("y_step", "0")))
         return cls(
             string=string,
             location=location,
